#perf: roll.rs
