+++ src/edges.rs
@@ -18,101 +18,11 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-use crate::{Edges, EdgesIntoIter, Label, Roll, RollIntoIter, ROLL_LIMIT};
+use crate::{Edges, EdgesIntoIter, Label, Roll};
 use serde::de::{MapAccess, Visitor};
 use serde::ser::SerializeMap;
 use serde::{Deserialize, Deserializer, Serialize, Serializer};
 use std::fmt::Formatter;
-
-impl<'a, K: Clone, V: Clone> Iterator for RollIntoIter<'a, K, V> {
-    type Item = (K, V);
-
-    #[inline]
-    fn next(&mut self) -> Option<Self::Item> {
-        while self.pos < ROLL_LIMIT {
-            if self.items[self.pos].is_some() {
-                let pair = self.items[self.pos].clone().unwrap();
-                self.pos += 1;
-                return Some(pair);
-            }
-            self.pos += 1;
-        }
-        None
-    }
-}
-
-impl<'a, K: Copy + PartialEq, V: Copy> IntoIterator for &'a Roll<K, V> {
-    type Item = (K, V);
-    type IntoIter = RollIntoIter<'a, K, V>;
-
-    #[inline]
-    fn into_iter(self) -> Self::IntoIter {
-        RollIntoIter {
-            pos: 0,
-            items: &self.items,
-        }
-    }
-}
-
-impl<K: Copy + PartialEq, V: Copy> Roll<K, V> {
-    #[must_use]
-    pub const fn new() -> Self {
-        Self {
-            items: [None; ROLL_LIMIT],
-        }
-    }
-
-    /// Make an iterator over all pairs.
-    #[inline]
-    pub const fn into_iter(&self) -> RollIntoIter<K, V> {
-        RollIntoIter {
-            pos: 0,
-            items: &self.items,
-        }
-    }
-
-    /// Is it empty?
-    #[inline]
-    pub fn is_empty(&self) -> bool {
-        self.len() == 0
-    }
-
-    /// Return the total number of pairs inside.
-    #[inline]
-    pub fn len(&self) -> usize {
-        let mut busy = 0;
-        for i in 0..ROLL_LIMIT {
-            if self.items[i].is_some() {
-                busy += 1;
-            }
-        }
-        busy
-    }
-
-    /// Insert a single pair into it.
-    ///
-    /// # Panics
-    ///
-    /// It may panic if you attempt to insert too many pairs.
-    #[inline]
-    pub fn insert(&mut self, k: K, v: V) {
-        for i in 0..ROLL_LIMIT {
-            if let Some((bk, _bv)) = self.items[i] {
-                if bk == k {
-                    self.items[i] = None;
-                    break;
-                }
-            }
-        }
-        for i in 0..ROLL_LIMIT {
-            if self.items[i].is_none() {
-                self.items[i] = Some((k, v));
-                return;
-            }
-        }
-        panic!("Out of space!")
-    }
-}
 
 impl Serialize for Edges {
     fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
@@ -201,67 +111,5 @@ fn serialize_and_deserialize() -> Result<()> {
     let bytes: Vec<u8> = serialize(&before)?;
     let after: Edges = deserialize(&bytes)?;
     assert_eq!(42, after.into_iter().next().unwrap().1);
-    Ok(())
-}
-
-#[test]
-fn insert_and_check_length() -> Result<()> {
-    let mut roll = Roll::new();
-    roll.insert(Label::Alpha(0), 42);
-    assert_eq!(1, roll.len());
-    roll.insert(Label::Alpha(1), 16);
-    assert_eq!(2, roll.len());
-    roll.insert(Label::Alpha(0), 16);
-    assert_eq!(2, roll.len());
-    Ok(())
-}
-
-#[test]
-fn empty_length() -> Result<()> {
-    let roll: Roll<u32, u32> = Roll::new();
-    assert_eq!(0, roll.len());
-    Ok(())
-}
-
-#[test]
-fn empty_iterator() -> Result<()> {
-    let roll: Roll<u32, u32> = Roll::new();
-    assert!(roll.into_iter().next().is_none());
-    Ok(())
-}
-
-#[test]
-fn insert_and_jump_over_next() -> Result<()> {
-    let mut roll = Roll::new();
-    roll.insert(Label::Alpha(0), 42);
-    let mut iter = roll.into_iter();
-    assert_eq!(42, iter.next().unwrap().1);
-    assert!(iter.next().is_none());
-    Ok(())
-}
-
-#[test]
-fn insert_and_iterate() -> Result<()> {
-    let mut roll = Roll::new();
-    roll.insert(Label::Alpha(0), 42);
-    roll.insert(Label::Alpha(1), 16);
-    let mut sum = 0;
-    for (_k, v) in roll.into_iter() {
-        sum += v;
-    }
-    assert_eq!(58, sum);
-    Ok(())
-}
-
-#[test]
-fn insert_and_into_iterate() -> Result<()> {
-    let mut roll = Roll::new();
-    roll.insert(Label::Alpha(0), 42);
-    roll.insert(Label::Alpha(1), 16);
-    let mut sum = 0;
-    for (_k, v) in roll.into_iter() {
-        sum += v;
-    }
-    assert_eq!(58, sum);
     Ok(())
 }
